requirements.txt
python-telegram-bot[job-queue]>=21.7,<22
requests>=2.32.3,<3
httpx>=0.27.2,<0.28
dateparser
python-dotenv

pytest>=8.3.3,<9
pytest-asyncio>=0.24.0,<0.25